requirements.txt
binary_reader==1.4.3
Pillow==9.5.0
xxhash==3.2.0
